+++ src/core/CnBaseItem.cpp
@@ -15,6 +15,11 @@ Cn::Access CnBaseItem::access() const
     return Cn::Access_Read;
 }
 
+const CnString &CnBaseItem::paramString() const
+{
+    return this->description();
+}
+
 void CnBaseItem::update(const CnVariant &value, Cn::StatusCode status, CnTimestamp timestamp)
 {
     // Note: base implementation do nothing